% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimations.R
\name{EM_likelihood_generic}
\alias{EM_likelihood_generic}
\title{Generic Euler-Maruyama Log-Likelihood for an SDE}
\usage{
EM_likelihood_generic(SDE_object, par, step_length, data)
}
\arguments{
\item{SDE_object}{An object of class \code{SDE} containing \code{drift}- and \code{diffusion} functions.}

\item{par}{Numeric vector. Parameter values to evaluate the likelihood in.}

\item{step_length}{Numeric vector. Real number that contains the temporal resolution of the trajectory used in the estimation.}

\item{data}{Numeric vector. Trajectory from the SDE model.}
}
\value{
Numeric. The negative log-likelihood of the observed data under the Euler-Maruyama approximation.
}
\description{
Computes the approximate log-likelihood of a one-dimensional SDE using
the Euler-Maruyama discretization.
}
\details{
This function can serve as a baseline for likelihood implementations of a given SDE.
}
