% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimations.R
\name{fit_SDE_EM}
\alias{fit_SDE_EM}
\title{Fit an SDE using Euler-Maruyama Maximum Likelihood}
\usage{
fit_SDE_EM(
  SDE_object,
  init_par,
  step_length = NA,
  data = NA,
  method = "BFGS",
  ...
)
}
\arguments{
\item{SDE_object}{An object of class \code{SDE} containing drift and diffusion and potentially data and step_length,}

\item{init_par}{Numeric vector. Initial guess for the parameters.}

\item{step_length}{Numeric vector. Optional temporal resolution of the trajectory in the data argument, if the SDE_object for instance does contain a trajectory.}

\item{data}{Numeric vector. Optional realized trajectory from the model given in the SDE_object, if this does not contain a trajectory, this can be passed explicitly here.}

\item{method}{Optimization method (default "BFGS").}

\item{...}{Additional arguments passed to \code{\link[stats]{optim}}.}
}
\value{
A list with elements \code{par} (estimated parameters) and \code{value} (negative log-likelihood).
}
\description{
Optimizes parameters of an SDE object using the Euler-Maruyama approximation.
}
