% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDE-class.R
\name{print.SDE}
\alias{print.SDE}
\title{Print an SDE}
\usage{
\method{print}{SDE}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SDE}.}

\item{...}{Additional arguments (currently ignored). Included for S3 consistency.}
}
\description{
Prints a summary of an object of class \code{SDE}.
}
