% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDE-class.R
\name{SDE}
\alias{SDE}
\title{Create a general SDE object}
\usage{
SDE(
  name,
  par,
  drift,
  diffusion,
  drift_dx = NULL,
  drift_dt = NULL,
  diffusion_dx = NULL,
  lamperti_transform = NULL,
  simulate_fun = NULL
)
}
\arguments{
\item{name}{Character. Name of the system that the Stochastic Differential Equation models.}

\item{par}{Numeric vector. Model parameters.}

\item{drift}{Function. Drift function drift(x, par).}

\item{diffusion}{Function. Diffusion function diffusion(x, par).}

\item{drift_dx}{Function. Optional derivative of drift with respect to x_t}

\item{drift_dt}{Function. Optional derivative of drift with respect to x_t. This is 0 for all autonomous processes.}

\item{diffusion_dx}{Function. Optional derivative of diffusion}

\item{lamperti_transform}{Function. Optional Lamperti transform that omits any parameters. I.e. the anti-derivative of 1 / diffusion(x)}

\item{simulate_fun}{Function. Optional custom simulation function.}
}
\value{
An object of class \code{SDE}.
}
\description{
Create a general SDE object
}
