% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_generic}
\alias{simulate_generic}
\title{Generic Euler-Maruyama scheme based Simulation for SDE-objects}
\usage{
simulate_generic(
  step_length,
  par,
  X_0,
  total_time,
  drift,
  diffusion,
  sample_method = "auto"
)
}
\arguments{
\item{step_length}{Numeric. Time increment for each simulation step.}

\item{par}{Numeric vector. Model parameters.}

\item{X_0}{Numeric. Initial value of the state variable.}

\item{total_time}{Numeric. Total simulation time.}

\item{drift}{Function. Drift function drift(x, par).}

\item{diffusion}{Function. Diffusion function diffusion(x, par).}

\item{sample_method}{Character. Method for generating random noise ("auto", "stats", "dqrng").}
}
\value{
A data.frame with columns \code{t} and \code{X}.
}
\description{
This function simulates a general one-dimensional SDE using the Euler-Maruyama method. It will be
used automatically if the SDE object does not provide a custom \code{simulate_fun}.
}
\keyword{internal}
