% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDE-class.R
\name{simulate_SDE}
\alias{simulate_SDE}
\title{Simulate an SDE}
\usage{
simulate_SDE(object, step_length, ...)
}
\arguments{
\item{object}{An object of class \code{SDE}.}

\item{step_length}{Numeric. Time increment for each simulation step.}

\item{...}{Additional arguments passed to the simulation function. For the generic simulation,
this must include:
\itemize{
\item \code{total_time}: Numeric. Total simulation time.
\item \code{X_0}: Numeric. Initial value of the state variable.
}
Additionally, \code{sample_method} can be specified for the generic simulation method.
Although this is optional. The valid parameters for this argument is ("auto", "stats", "dqrng"). Default is "auto".}
}
\description{
Simulates the trajectory of an SDE object.
}
\details{
Depending on whether or not a custom simulation method has been given to \code{simulate_fun}. This method will either:
\enumerate{
\item Use the custom simulation function \code{simulate_fun} of the \code{SDE}-object and sample according to it, or
\item Use a generic Euler-Maruyama simulation based on the \code{drift}- and \code{diffusion}-functions from the \code{SDE}-object,
in which case the arguments: \code{step_length}, \code{total_time}, and \code{X_0} \strong{must} be provided in \code{...}.
}
}
