% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runge_kutta_vect.R
\name{runge_kutta_vect}
\alias{runge_kutta_vect}
\title{Vectorized Runge-Kutta Method (RK4)}
\usage{
runge_kutta_vect(x0, y0, h, func, n)
}
\arguments{
\item{x0}{A numeric vector of initial values for the independent variable.}

\item{y0}{A numeric vector of initial values for the dependent variable.}

\item{h}{A numeric scalar that defines the step size for the Runge-Kutta method.}

\item{func}{A function that describes the derivative in the point some point \verb{(x,y)}. It should take \code{x0} and \code{y0} as arguments and be vectorized}

\item{n}{An integer representing the number of sub-steps to iterate through. Setting n = 1 is typically sufficient in estimation applications.}
}
\value{
A numeric vector of estimated values for the function described \code{func}, i.e. the ODE, evaluated at the points \code{x0 + h}.
The output will have the same length as \code{x0} and \code{y0}.
}
\description{
This function implements the \strong{Runge-Kutta 4th Order (RK4)} method for solving
ordinary differential equations (ODEs). It is designed to handle \strong{vectorized}
input for initial conditions \code{x0} and \code{y0}. The function gives the approximation
to the solution of the initial value problem at \code{x0 + h} for all starting positions \code{x0}
based on the provided function, \code{func}, a step-size \code{h}, and number of sub-steps \code{n}.
}
\examples{
# Example of using runge_kutta_vect
func <- function(x, y) { x + y }  # Simple ODE: y'(x) = x + y
x0 <- c(0, 1)  # Initial condition for x
y0 <- c(1, 2)  # Initial condition for y
h <- 0.1  # Step size
n <- 1  # Number of sub-steps
result <- runge_kutta_vect(x0, y0, h, func, n)
print(result)
}
